requirements.txt
Django==1.11.13
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
waitress==1.1.0
